added a dialog fragment as a form  before and after filling the fuel

+++ app/src/main/java/com/fourofourfound/aims_delivery/database/daos/Daos.kt
@@ -33,7 +33,7 @@ interface TrailerDao {
     suspend fun insertTrailer(trailer: DatabaseTrailer)
 
     @Query("update DatabaseTrailer set fuelQuantity =:fuelQuantity where trailerId =:trailerId")
-    suspend fun updateTrailerFuel(trailerId: Int, fuelQuantity: Int)
+    suspend fun updateTrailerFuel(trailerId: Int, fuelQuantity: Double)
 }
 
 @Dao
