+++ app/src/main/java/com/fourofourfound/aims_delivery/database/relations/TripWithInfo.kt
@@ -99,7 +99,7 @@ fun List<TripWithInfo>.asDomainModel(): List<Trip> {
                 )
             }
             destination.trailer.apply {
-                trailerInfo = TrailerInfo(trailerId, trailerCode, trailerDesc, fuelQuantity)
+                trailerInfo = TrailerInfo(trailerId, trailerCode, trailerDesc, fuelQuantity.toDouble())
             }
 
             destination.truck.apply {
